Refactored crypto keys and moved non-test parts to outside the tests
