+++ crypto/signature/ed25519.go
@@ -2,18 +2,17 @@ package signature
 
 import (
 	"fmt"
+	"github.com/orbs-network/orbs-network-go/crypto/keys"
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
 	"golang.org/x/crypto/ed25519"
 )
 
 const (
-	ED25519_PUBLIC_KEY_SIZE_BYTES  = 32
-	ED25519_PRIVATE_KEY_SIZE_BYTES = 64
-	ED25519_SIGNATURE_SIZE_BYTES   = 64
+	ED25519_SIGNATURE_SIZE_BYTES = 64
 )
 
 func SignEd25519(privateKey primitives.Ed25519PrivateKey, data []byte) (primitives.Ed25519Sig, error) {
-	if len(privateKey) != ED25519_PRIVATE_KEY_SIZE_BYTES {
+	if len(privateKey) != keys.ED25519_PRIVATE_KEY_SIZE_BYTES {
 		return nil, fmt.Errorf("cannot sign with ed25519, private key invalid")
 	}
 	signedData := ed25519.Sign([]byte(privateKey), data)
@@ -21,7 +20,7 @@ func SignEd25519(privateKey primitives.Ed25519PrivateKey, data []byte) (primitiv
 }
 
 func VerifyEd25519(publicKey primitives.Ed25519PublicKey, data []byte, signature primitives.Ed25519Sig) bool {
-	if len(publicKey) != ED25519_PUBLIC_KEY_SIZE_BYTES {
+	if len(publicKey) != keys.ED25519_PUBLIC_KEY_SIZE_BYTES {
 		return false
 	}
 	return ed25519.Verify([]byte(publicKey), data, signature)
